Test made, package updated, fixed spacing
Fixed the spacing in school's page when logged out, added functionality to log in on school's page which should've been there but I forgot to add it, added the tester to school, updated package.json accordingly. Also, schools.json is missing the declaration of Jane being a teacher, so I added that.